requirements.txt
python-dotenv == 1.0.1
snowflake-connector-python == 3.12.2